backend/requirements.txt
django~=5.2
djangorestframework~=3.17
python-dotenv~=1.0
psycopg2-binary~=2.9
